refactor: change structura
