+++ src/multi.ts
@@ -20,7 +20,7 @@ if (numCpus > 1) {
     }
     cluster.on("exit", (worker, code, signal) => {
       console.log(`worker ${worker.process.pid} died`);
-      // cluster.fork();
+      cluster.fork();
     });
   } else {
     startServer();
