+++ src/tests/utils.test.ts
@@ -1,4 +1,4 @@
-import * as utils from "../utils";
+import * as utils from "../utils/utils";
 
 const id = "123456789";
