+++ src/utils/reqBody.ts
@@ -1,0 +1,19 @@
+import { IncomingMessage } from "http";
+
+export const getBody = (req: IncomingMessage): Promise<string> => {
+  return new Promise((resolve, reject) => {
+    try {
+      let body = "";
+
+      req.on("data", (chunk) => {
+        body += chunk.toString();
+      });
+
+      req.on("end", () => {
+        resolve(body);
+      });
+    } catch (error) {
+      reject(error);
+    }
+  });
+};
